+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 - Resolved #910: Remember last site user visited.
 - Resolved #851: Add Auto-Save mode for posts.
+- Fixed #918: Fix commenting enabled checkbox default state.
+- Resolved #924: Close comment modal after reply
+- Resolved #928: Fix loadmore button for file selection modal.
 
 
 2018.7.2